ENH: Check for open orders in order_value
Keyword arg, ignore_open_orders, added to the order_value api method.

If ignore_open_orders is False, the order will not be submitted if
there are any open orders for that particular sid.
Open orders are ignored otherwise.

This is to help keep order_value from placing extra, or mis-sized
orders when there are open orders out already.