ENH: time in force models DISCUSSION ONLY
Time in force models represent when an order is to be active. I modeled
them after the Execution styles and made the default behavior the same
as IB’s. The idea is to support more IB native order types, these get
passed to TWS via Order.m_tiff.

Classes:
TimeInForceModel: abstract base class

DayOrder:
IB’s default order type, will be cancelled if not filled by EOD.

GoodTillCancelled:
Essentially what orders do in zipline now, the order will remain on the
blotter until it fills or is explicitly cancelled.

ImmediateOrCancel:
Unfinished as of now. Any portion of the order that does not get filled
immediately will be cancelled.

GoodTillDate:
Order will be good until a specified date. It will be cancelled after
the good_till_date

 GoodBetweenDates:
This is not a native time in force model in IB, but they do have fields
for ‘good before/after’ dates so I added this one as well.