ENH: time in force added to TradingAlgorithm DISCUSSION
Time in force models were added to the TradingAlgorithm order method.

An order can specify a time in force model by passing an instance of a
TimeInForceModel, or a string id can be used to specify the model.

The default time in force is a DayOrder that is cancelled at EOD if it
doesn’t fill.