BUG: Fixed issue in tests causing incorrect failures.
The test architecture was set up that self.initialized was never
set to true. This caused my new order validation code to fail incorrectly.
The tests have been modified to set initialized to true. Also,
there was an illegal call to set_slippage after init that passed
bacause initialized was never set to true. A kwarg was added to
the test init method that allowed slippage to be passed.