ENH: Added EventManager to zipline.utils
Added an EventManager class to simplify handling events that do not
occur on every bar.

This version works for specifying the number of trading days between
events and limiting the number of times per day it can occur.

A rule function needs to be passed to the EventManager when it is
initialized, which will be called for a final entry decision when all
other criteria has been met. Some common case rule functions will be
added in another commit.

This class will be called on every bar, possibly several times, so
speed is a concern. Please share any suggestions to boost performance.