ENH: Added Support for multiple events/rules
I changed the old EventManager to PeriodicEvent because it does well
for handling a single event. I also tightened up the scope of what this
is for. It only accepts a datetime arg, and only returns a boolean
value.

I made a new EventManager that associates a named event with a rule. It
creates a new attribute for each rule, keyed by the name. Calling
manager.event_name(dt) returns the decision for that rule.