ENH: Support for scheduled function calls
Added zipline.utils.events.

EventOffset class
The EventOffset class is to give people the ability to schedule
events to occur at frequencies other than the backtest
frequency. It increments an internal clock by a pandas
offset frequency and calls the functions associated with it
when the offset has been satisfied.

If the frequency is longer than a trading day, an optional
time_func can be passed to specify when in the day the
event can occur, it returns True by default.

EventContainer class
The TradingAlgorithm class is given an EventContainer
attribute that EventOffsets can be added to. This class
calls the EventOffset objects before the remainder of the
handle_data loop.