ENH: Forward *args and **kwargs from __init__ to initialize.
The reason for this change is that it was impossible to
use a TradingAlgorithm that passes in an initialize function
in a parameter optimization. This issue is discussed in #455.

It seems easy enough to just forward any left-over *args
and **kwargs to the user-defined initialize function. As these
are only passed when they are defined we do not break backwards
compatibility with this.