MAINT Increase default volume from 1000 to 1e9
For more discussion see https://github.com/quantopian/zipline/pull/485.

Basically, 1000 is just a number that was supposed to be high enough if no volume was available. It turns out that number is actually very low so now we are increasing it so that volume restrictions should no matter. 1e9 of shares ought to be enough for anybody ;).

Thanks to @jlowin for pointing that out.