PERF: Reduce memory usage during risk metric calculations.
Instead of passing a pd.Series to calculations, use a memory view of the
underlying numpy array which is sliced to the exact dt, to prevent
calculation like beta creating some temporary copies of the Series.

The creation of the temporary Series was consuming more memory than
needed, with the index being the high penalty.