MAINT: Use sqlite parameterization instead of string formatting.
For security best practices, though it is unlikely that a injectable
value would get to this point, use the Python sqlite3 module's
parameterization for providing parameters.

String formatting is still used on the template for providing the table
name and the number of parameter tokens to use for the IN statement,
since those values are not parameterizable. Those format values there
are hardcoded values, i.e. `splits`, `mergers`, `dividends` and the `?`
character.

Also, factor out the shared query logic between splits, mergers, and
dividends, since this change would have changed all three places in the
same way.