[BUG] For Discussion- Fixes risky risk free rate
Our default risk-free-rate is the 10-year yield, which is
considered risky by 99% of investors. 

This is a quick fix to make the default the 3-month T-bill rate, 
which is a widely used risk-free instrument. 

An api function to set the risk free rate within initialize might be a 
worthwhile project, but just changing this one string solves the
bulk of the issue.