MAINT: Remove consume identifiers from asset finder.
Change use of the identifier keyword and methods for the asset finder to
use insert_metadata instead.

Many cases of the identifier keyword removal are now covered by the
algorithm run and various sources now using insert_metadata.

The main impetus of this patch is to reduce the surface area of ways
that assets are added to the asset finder in anticipation of porting the
asset finder over to be sqlite backed.

This patch does require some calls to populate_cache to be added, but
those will be removed during the aforementioned port.