MAINT: Change asset finder map index to an explicit populate.
The map_identifier_index_to_sids was both returning the matched sids and
creating assets for sids that did not yet exist.

Both change the name to explicitly show that it populates, and move it
out of the AssetFinder since it is more of a convenience function for
DataFrame and the like backed algos than it is core functionality of the
AssetFinder.