Only access position values once during stat calc.
Instead of calculating the position values for each stat result, e.g.
gross_exposure, net_liquidity etc. get the positions upfront and then
calculate the period and position stats in order, passing each value
explicitly to the ones that follow it in the dependency chain.

On the path for wiring up DataPortal, so that the upkeep of
last_sale_price can be removed in favor accessing those values on demand.