MAINT: Remove position tracker from perf period.
Instead of each performance period holding a reference to the same
position tracker, have the performance tracker pass the position tracker
explicitly to the periods when it is needed for performance
calculations.

Also, remove the position tracker from serialization of perf
periods (now only the perf tracker tracks the position tracker), which
is a side benefit.

Main intent is to work towards the path where the position stats are
calculated once by the performance tracker and then passed to the
methods which now take the position tracker.