MAINT: Pass data to perf period calcs from tracker
Move the calculation of position stats and perfomance period
calculations up the stack so that the performance tracker controls the
timing (and frequency) of snapshotting the position and performance
periods' data.

This intent of this change is to prepare the tracker and related classes
so that a dt can be passed explicitly during calculations when attached
to the incoming data portal changes.

Should have some slight performance benefits since required calculations
are done closer to once a bar.

This patch also includes changes to how the algorithm object caches the
portfolio and account objects. They are changed here instead of
separately because the `update_portfolio` and related flags were
removed.