MAINT: Rename contract multiplier to contract size
We have been using the term 'contract multipler' in our
futures logic, but we should be using 'contract size'.
The two are related as follows:

contract multiplier = contract size * tick

where tick is the minimum price fluctuation of a single
unit (e.g. a single barrel of oil).

So, for WTI Crude Oil (CL), the above equation is:

$10.00 = 1000 barrels * 0.01 cents