REF: Use daily bar reader spot price in portal
On the path of removing the need to hardcode/or pass around a 'first
trading day' value, use the calendar alreday embedded in the bcolz file
by the daily bar writer by using the daily bar reader's spot_price
method.

Removes one, but not all of the uses of the `FIRST_TRADING_DAY` value.