ENH: Remove hard coding of first day in portal
Use daily bar reader for daily history values in data portal, which
removes the need maintaining the first trading day in the data portal
class. The daily bar reader reads the calendar and asset starts from the
data written to the daily bar file. Using that data removes the need to
hard code the values in the data portal module, without having to pass
the values down to the data portal as parameters.

Removes daily pricing adjustment from data portal since that value is
handled by the daily bar reader.

Also, update how daily benchmarks (which uses the data_portal) handles
the case where the daily bar reader does not have data for the day
before the start of the backtest. (Since the reader throws an exception
when data is accessed before the calendar start.)