ENH: Allow configurable equity minute data start.
Instead of using a hardcoded start date for the calculations of the
index into the equity minute data, use a metadata file written by the
equity minute writer to set the first trading day of the equity minute
set.

Intended to allow the data portal to be used with datasets with
different ranges. (i.e. not just the start date of 2002-01-02)