REF: Remove data offset calc from clock.
Remove the responsibility of calculating the minute dataset index from
the clock, i.e. the offset that is used for the 'fast' lookup path in
the bcolz data. Instead, the clock calculates all the minutes that the
clock will run through and provides a map of days to the minutes for
each day, and a map of minutes to which day the minute occurs on.

Those two maps are then used by the data_portal to do on demand
calculations of the dataset offset, instead of the clock calculating
the value every minute.

This is on the path to removing all hardcodings of the minute offsets,
by removing the hardcoding from the clock, with the data_portal taking
responsibility for the offsets (work that follows this will remove the
hardcoding from the data_portal.)