REF: Pass minute reader instead of path to portal
Start down path of abstracting out the reader, so that other readers,
e.g. a DataFrame can be implemented. (Those implementations would
require more of the read logic to be moved from the DataPortal to the
reader.)

This is also required for an internal override of the start date on the
bcolz reader, which is why this change is so minimal.