PERF: Fix runtime regression on daily windows.
Re-using the pipeline infrastructure resulting in a bottleneck when
using a small number of stocks, where the read of the entire numpy
vector when only one stock was used was prohibitive.

Instead, base a solution off of @jbredeche's original solution, but
place the logic (at least the indexing logic into the DailyBcolzReader.)

We have an eye on making this eventually share code with pipeline, but
using the simpler non-shared solution for now.