ENH: Add initial commit for DataPortal and readers
Moved from the `lazy-mainline` branch,
https://github.com/quantopian/zipline/pull/858

With the removal of methods that are not covered by the test_dataportal
suite.

The intent of this patch to provide the basic class and readers
developed on that branch, so that the use of creating the object and
opening paths etc. can be tested internally.

Important to note that the only usage is in the test_dataportal suite,
before any usage of the data_portal by the tradesimulation, a writer
will should be provided for publicly available data.

co-author: @jbredeche <jean@quantopian.com>