REF: Add futures reader to data portal.
Instead of passing a path to the futures data, provide a reader, to
align with the usage of the equity minute reader.

Also allow an argument of the future_daily_reader, though no such reader
yet exists.

It may be that future and equity readers share an interface, and a
further improvement would be providing an abstract base class.