REF: Require data_frequency as spot_value param
So that the data_portal is not tied to solely daily or minute data
frequencies, make the get_spot_value method require a data frequency.

This change required that more classes held a reference to the
simulation's data frequency, e.g. the blotter and performance trackers.

Update serialization to reflect those changes.