ENH: Provide method for last trade dt.
Add a data_portal method `get_last_traded_dt` which the dt passed to the
method if there is a trade on that bar, or the previous dt with a trade
if there is no trade on the current dt.

This method will be used to proved a `.dt` lookup on SidView, but the
connection to the SidView is not made with this patch.

To implement the get_last_traded_dt a corresponding method
was added to the minute bar reader, with a supporting DatetimeIndex
which matches the data sets indices. Also, change
`_find_position_of_minute` to use the index, since the lookup is faster
than calculating a Timestamp and doing the offset calculation every
time. (As seen by the improvement in the ffill tests in test_dataportal.)

Also, change history tests to use a shared data_portal to avoid setup
cost of repeatedly creating the lookup table in the reader.