ENH: Forward fill a history window with prior data
The forward-filling logic for price left leading nans in the window,
even if a value existed before the start of the window. Improve the
forward filling behavior by getting the last value before window and
setting it as the first element before forward filling.

Add tests for equities and futures. The equities tests extend over a
split, to ensure that the window is fully adjusted.